Move `FakeUIRefreshControl` to `FeedViewController` test helper extension
