+++ EssentialFeed/EssentialFeediOSTests/Feed UI/Controllers/FeedViewControllerTests.swift
@@ -258,18 +258,3 @@ final class FeedViewControllerTests: XCTestCase {
     }
     
 }
-
-class FakeUIRefreshControl: UIRefreshControl {
-    var _isRefresing: Bool = false
-    
-    override var isRefreshing: Bool { _isRefresing }
-    
-    override func beginRefreshing() {
-        _isRefresing = true
-    }
-    
-    override func endRefreshing() {
-        _isRefresing = false
-    }
-    
-}
