--- EssentialFeed/EssentialFeediOSTests/Helpers/UIButton+TestHelpers.swift
@@ -1,14 +1,0 @@
-//
-//  UIButton+TestHelpers.swift
-//  EssentialFeediOSTests
-//
-//  Created by Prabhat Tiwari on 08/07/25.
-//
-
-import UIKit
-
-extension UIButton {
-    func simulateTap() {
-        simulate(event: .touchUpInside)
-    }
-}
