--- EssentialFeed/EssentialFeediOSTests/Helpers/UIRefreshControl+TestHelpers.swift
@@ -1,14 +1,0 @@
-//
-//  UIRefreshControl+TestHelpers.swift
-//  EssentialFeediOSTests
-//
-//  Created by Prabhat Tiwari on 08/07/25.
-//
-
-import UIKit
-
-extension UIRefreshControl {
-    func simulatePullToRefresh() {
-        simulate(event: .valueChanged)
-    }
-}
